Seperate resolvers from main
